src/PFAN-SurgiATM-PGSA/requirements.txt
addict==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
archspec==0.2.3
boltons==24.0.0
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.3.2
colorama==0.4.6
conda==24.11.2
conda-libmamba-solver==24.9.0
conda-package-handling==2.4.0
conda_package_streaming==0.11.0
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
dominate==2.9.1
einops==0.8.2
filelock==3.25.2
fonttools==4.62.1
frozendict==2.4.6
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
h2==4.1.0
hf-xet==1.4.3
hpack==4.0.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.9.0
hyperframe==6.0.1
idna==3.10
ImageIO==2.37.3
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.5.0
lazy-loader==0.5
libmambapy==1.5.12
mamba==1.5.12
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
menuinst==2.2.0
mmcv==2.2.0
mmengine==0.10.7
mpmath==1.3.0
networkx==3.6.1
numpy==1.26.4
opencv-python==4.13.0.92
packaging==24.2
pillow==12.1.1
pip==24.3.1
platformdirs==4.3.6
pluggy==1.5.0
protobuf==6.33.6
pycosat==0.6.6
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
pyparsing==3.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.3
rich==14.3.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.8
safetensors==0.7.0
scikit-image==0.26.0
scipy==1.17.1
sentry-sdk==2.57.0
setuptools==65.6.3
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sympy==1.14.0
termcolor==3.3.0
tifffile==2026.3.3
timm==1.0.26
torch==2.1.0+cu121
torchvision==0.16.0+cu121
tornado==6.5.5
tqdm==4.67.1
triton==2.1.0
truststore==0.10.0
typer==0.24.1
typing_extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.3.0
visdom==0.2.4
wandb==0.25.1
websocket-client==1.9.0
wheel==0.45.1
yapf==0.43.0
zstandard==0.23.0
